+++ src/main/java/dansplugins/rpsystem/commands/BirdCommand.java
@@ -1,7 +1,7 @@
 package dansplugins.rpsystem.commands;
 
 import dansplugins.rpsystem.MedievalRoleplayEngine;
-import dansplugins.rpsystem.Messenger;
+import dansplugins.rpsystem.utils.Messenger;
 import dansplugins.rpsystem.data.EphemeralData;
 import dansplugins.rpsystem.integrators.MailboxesIntegrator;
 import dansplugins.rpsystem.utils.ColorChecker;
