+++ src/main/java/dansplugins/rpsystem/commands/EmoteCommand.java
@@ -1,7 +1,7 @@
 package dansplugins.rpsystem.commands;
 
 import dansplugins.rpsystem.MedievalRoleplayEngine;
-import dansplugins.rpsystem.Messenger;
+import dansplugins.rpsystem.utils.Messenger;
 import dansplugins.rpsystem.data.PersistentData;
 import dansplugins.rpsystem.utils.ColorChecker;
 import org.bukkit.ChatColor;
