+++ src/main/java/dansplugins/rpsystem/eventhandlers/ChatHandler.java
@@ -2,7 +2,7 @@ package dansplugins.rpsystem.eventhandlers;
 
 import dansplugins.factionsystem.externalapi.MF_Faction;
 import dansplugins.rpsystem.MedievalRoleplayEngine;
-import dansplugins.rpsystem.Messenger;
+import dansplugins.rpsystem.utils.Messenger;
 import dansplugins.rpsystem.data.EphemeralData;
 import dansplugins.rpsystem.data.PersistentData;
 import dansplugins.rpsystem.integrators.MedievalFactionsIntegrator;
